refactored the banner models to be in the room
screen view model
