+++ ElementX/Sources/Screens/RoomScreen/RoomScreenViewModelProtocol.swift
@@ -20,4 +20,6 @@ import Foundation
 protocol RoomScreenViewModelProtocol {
     var actions: AnyPublisher<RoomScreenViewModelAction, Never> { get }
     var context: RoomScreenViewModel.Context { get }
+    
+    func timelineHasScrolled(direction: ScrollDirection)
 }
